requirements.txt
Django==4.2
dj-database-url==2.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
whitenoise==6.5.0
django-heroku==0.3.1
requests==2.31.0
bardapi==0.1.11


